% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfam-main.R
\name{cfam}
\alias{cfam}
\title{A Constrained functional additive model (main function)}
\usage{
cfam(y, Tr, X1D = NULL, Xsc = NULL, nbasis.t = NULL,
  rho.grid = c(0.1, 0.3, 0.5), eps = 0.05, max.iter = 20,
  cd.eps = 0.005, cd.max.iter = 100, type = "AIC",
  lambda.grid = seq(0.05, 0.25, length.out = 10), main.effect = FALSE,
  lambda.grid2 = seq(0.1, 0.5, length.out = 10), warm.start = TRUE,
  warm.start.rho = 0.1, warm.start.lam = 0.15, csim.ini = TRUE,
  linear.link = NULL, trace = FALSE, plots = TRUE, scale.X = TRUE,
  sparse.coef = FALSE, pve = 0.95)
}
\arguments{
\item{y}{treatment outcomes, n-by-1 vector}

\item{Tr}{treatment indicators, n-by-1 vector; each element represents one of the K available treatment options}

\item{X1D}{a list of p pretreatment functional regressors; each of the functional regressors is evaluated on a grid (in [0,1]), giving a matrix (or a data.frame)}

\item{Xsc}{a matrix of q pretreatment scalar-regressors, n-by-q matrix (or a n-by-q data.frame)}

\item{nbasis.t}{a length K+1 vector; each element specifies the number of B-spline basis funtions for approximating the treatment-specific link function; the last element is for the "main effect" link function; the default is \code{nbasis.t=NULL}, and will be determined depending on the sample size.}

\item{rho.grid}{a grid vector of (ridge-type) smoothing parameters the p-splines in the link function estimation}

\item{eps}{a value specifying the convergence criterion of algorithm for estimating coef (linear prjections associated with the functional regressors)}

\item{max.iter}{an integer value specifying the maximum number of iterations for estimating coef (linear prjections associated with the functional regressors)}

\item{cd.eps}{a value specifying the convergence criterion of coordinate descent algorithm for estimating the link functions}

\item{cd.max.iter}{an integer value specifying the maximum number of coordinate descent iterations for estimating the link funtions}

\item{type}{can choose bewteen \code{"AIC"}, \code{"BIC"}, and \code{"GCV"}, for the sparsity tuninig parameter selection.}

\item{lambda.grid}{a grid of the sparsity tuning parameters, \code{lambda}, used in the \eqn{L1} regularization.}

\item{main.effect}{if \code{TRUE}, first regress out the main effect of the regressors on the outcome, fitted by \code{fam}; the default is \code{FALSE}.}

\item{lambda.grid2}{when \code{main.effect=TRUE}, a grid of the sparsity tuning parameters to be used for the main effect estimation}

\item{warm.start}{if \code{TRUE}, fit a \code{cmim} with some given penalty parameters to obtain a reasonable starting value for the index coeffiicent vectors.}

\item{warm.start.rho}{a smoothness penalty to be used in obtaining an initial estimate.}

\item{warm.start.lam}{a sparsity penalty to be used in obtaining an initial estimate.}

\item{csim.ini}{if \code{TRUE}, use a \code{csim} (a constrained single-index model) estimate as a warm start initial estimate  of the index coeffiicent vectors.}

\item{linear.link}{a length (p+q) vector of 1's or 0's, indicating whether to restrict the jth link functions to be linear or not; if \code{NULL}, will take the vector of 0's.}

\item{trace}{if \code{TRUE}, show the trace of the fitting procedure; the default is \code{FALSE}.}

\item{plots}{if \code{TRUE}, produce plots of the estiamated link functions and the estiamated coefficient functions}

\item{scale.X}{if \code{TRUE}, scale the FPC scores to have a unit variance for the parameter estimation (and will be scaled back in the final estimate)}

\item{sparse.coef}{if \code{TRUE}, estimate spasre coef vectors associated with the FPC scores, using lasso with \code{glmnet}; the default is \code{FALSE}.}

\item{pve}{the proportion of variance explained in determining the number of FPCs for the functional regressors}
}
\value{
a list of information of the fitted model including
\item{cmim.obj}{a \code{cmim} (a constrained multiple-index model) object containing information about the fitted interaction-effect model, based on the FPC (scalar) scores and the scalar regressors.}
\item{mim.obj}{a \code{mim} (a multiple-index model) object containing information about the fitted main-effect model, based on the FPC (scalar) scores and the scalar regressors, if \code{main.effect=TRUE}.}
\item{scale.param}{a length (p+q) vector representing the scale parameters associated with the (p+q) regression component functions; 0 indicates the associated component function is shrinked to 0 (due to the L1 regularization).}
\item{scale.param.path}{a path of \code{scale.param} over a decreasing sequence of the sparsity parameters, \code{lambda}.}
\item{lambda.opt}{a value indicating the estimated optimal sparsity parameter, \code{lambda}.}
\item{coef}{a length p list of the estimated coef (the linear projection coefficients) associated with the FPC scores.}
\item{link.fn.fit}{a list of information about the estimated (p+q) link functions, including the knot sequences used in the B-spline approximations and the smoother matrices.}
\item{beta.t.coef}{a list of the estimated B-spline coefficient vectors associated with the (p+q) link functions.}
\item{beta.0.coef}{a list of the estimated B-spline coefficient vectors associated with the (p+q) link functions, used in orthogonalzing the link functions against the main-effect.}
\item{coef.fn}{a length p list of the estimated coefficient functions associated with the unctional regressors.}
\item{eigen.fn}{a length p list of the matrices of eigenfunctions associated with the p functional regressors.}
\item{mean.fn}{a length p list of the mean functions associated with the p functional regressors.}
\item{rho.opt}{a length (p+q) vector of the estimated optimal penalty parameters used in the P-splines, associated with the (p+q) link functions}
\item{link.fn.plot}{a plot object for the (p+q) link functions.}
\item{coef.fn.plot}{a plot object for the p coefficient functions.}
}
\description{
\code{cfam} is the main function for fitting the constrained functional additive model (CFAM).
\code{cfam} uses pre-treatment functional regressors, X1D, and scalar regressors, Xsc, for modeling a scalar-valued treatment outcome y.
For each of the functional regressors, \code{cfam} reduces its dimension via a data-driven linear projection to define a scalar-valued index variable,
and fits a treatment-specific additive model over those index variables and the scalar regressors simultaneousely.
The differential treatment effect is estimated by these treatment-specific nonparametrically-defined link functions on the indices and the scalar regressors.
For simultaneous variable selection for the treatment effect modifiers,
 \code{cfam} estimates a sparse combination of the components of the additive model via a \eqn{L1} regularization.
The resulting \code{cfam} object can be used to estimate an optimal treatment selection rule
for a new patient with pre-treatment clinical information.
}
\details{
The sequence of the model coefficients implied by the tuning paramters \code{lambda} is fit by a coordinate descent algorithm.
}
\examples{
## generate a dataset
dat <- dataGnFn(n=400, p=10, q=3, contrast ="nonlinear")
y <- dat$y   # a vector of scalar-valued outcomes
Tr <- dat$Tr  # a vector of treatment indicators
X1D <- dat$X  # a list of 1-D functional regressors
Xsc <- dat$Z  # a matrix of scalar regressors

# a grid of tuning parameters for the treatment effect modifier selection
lambda.grid <- seq(0.1, 0.25, length.out = 10);
cfam.obj <- cfam(y=y, Tr=Tr, X1D=X1D, Xsc=Xsc,lambda.grid=lambda.grid, type="AIC",
nbasis.t= c(5,5,7), main.effect = FALSE)
cfam.obj$lambda.opt
#cfam.obj$coef  # coefficient associated with the functoinal covariates
# a length (p+q) vector of the shrinkage factors:
cfam.obj$scale.param  # 0 indicates that the corresponding is not selected as a treatment effect modifier
#cfam.obj$scale.param.path
cfam.obj$link.fn.plot[[1]]
cfam.obj$coef.fn.plot[[1]]

# can also fit FAM to model the covariates' main effects;
# due to the orthogonality between the main and interaction effect components,
# the main effect model can be fitted seprately from the interaction effect model,
# and regressed out before fitting the interaction effect model.

# a grid of tuning parameters for variable selection for the main effect model;
# this is to avoid over-fitting
lambda.grid2 <- seq(0.3, 0.6, length.out = 10);
cfam.obj2 <- cfam(y=y, Tr=Tr, X1D=X1D, Xsc=Xsc, lambda.grid=lambda.grid,
nbasis.t= c(5,5,7), main.effect = TRUE, lambda.grid2 =lambda.grid2)
cfam.obj2$lambda.opt
cfam.obj2$scale.param  # can identify the selected treatment effect modifiers
cfam.obj2$link.fn.plot[[1]]
cfam.obj2$coef.fn.plot[[1]]
cfam.obj2$mim.obj$scale.param  # can also identify the selected main-effect covariates
cfam.obj2$mim.obj$link.fn.plot[[1]]
}
\seealso{
\code{pred.cfam},  \code{cmim},  \code{fam}
}
\author{
Park, Petkova, Tarpey, Ogden
}
