% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfam-main.R
\name{cmim}
\alias{cmim}
\title{A Constrained multiple-index model, for estimating interaction effects}
\usage{
cmim(y, Tr, X.list, nbasis.t = NULL, rho.grid = c(0.1, 0.3, 0.5),
  lambda.grid = seq(0.1, 0.3, length.out = 10), type = "AIC",
  coef = NULL, csim.ini = FALSE, cd.eps = 0.001, cd.max.iter = 100,
  linear.link = NULL, eps = 0.01, max.iter = 20, trace = FALSE,
  scale.X = TRUE, sparse.coef = TRUE, plots = TRUE)
}
\arguments{
\item{y}{treatment outcomes, n-by-1 vector}

\item{Tr}{treatment indicators, n-by-1 vector; each element represents one of the K available treatment options}

\item{X.list}{a list of V matrices, in which each matrix is the data matrix corresponding to each of the V groups of the scalar-valued covariates.}

\item{nbasis.t}{a length K+1 vector; each element specifies the number of B-spline basis funtions for approximating the treatment-specific link function; the last element is for the "main effect" link function; the default is \code{nbasis.t=NULL}, and will be determined depending on the sample size.}

\item{rho.grid}{a grid vector of (ridge-type) smoothing parameters the P-splines in the link function estimation}

\item{lambda.grid}{a grid of the sparsity tuning parameters, \code{lambda}, used in the \eqn{L1} regularization.}

\item{type}{can choose bewteen \code{"AIC"}, \code{"BIC"}, and \code{"GCV"}, for the sparsity tuninig parameter selection.}

\item{coef}{a list of length p; can explicitely specify an initial estimate of the index coef vectors; the default is \code{NULL}.}

\item{csim.ini}{if \code{TRUE} and \code{coef} is not explictely given, use a \code{csim} (a constrained single-index model) estimate as an initial estimate of the index coefficient vectors.}

\item{cd.eps}{a value specifying the convergence criterion of coordinate descent algorithm for estimating the link functions}

\item{cd.max.iter}{an integer value specifying the maximum number of coordinate descent iterations for estimating the link funtions}

\item{linear.link}{a length V vector of 1's or 0's, indicating whether to restrict the jth link functions to be linear or not; if \code{NULL}, will take the vector of 0's.}

\item{eps}{a value specifying the convergence criterion of algorithm for estimating the index coefficient vectors}

\item{max.iter}{an integer value specifying the maximum number of iterations for estimating the index coefficient vectors}

\item{trace}{if \code{TRUE}, show the trace of the fitting procedure; the default is \code{FALSE}.}

\item{scale.X}{if \code{TRUE}, scale the covariates to have a unit variance for the parameter estimation (and will be scaled back for the final estimate)}

\item{sparse.coef}{if \code{TRUE}, estimate spasre index coefficient vectors, using lasso with \code{glmnet}; the default is \code{FALSE}.}

\item{plots}{if \code{TRUE}, produce plots of the estiamated link functions and the estiamated coefficient functions}
}
\value{
a list of information of the fitted constrained multiple-index model including
\item{coef}{a length V list of the estimated index coefficient vectors}
\item{link.fn.fit}{a list of information about the estimated V link functions, including the knot sequences used in the B-spline approximations and the smoothers}
\item{scale.param}{a length V vector representing the scale parameters associated with the V regression component functions; 0 indicates the associated component function is shrinked to 0 (due to the L1 regularization).}
\item{scale.param.path}{a path of \code{scale.param} over a decreasing sequence of the sparsity parameters, \code{lambda}.}
\item{lambda.opt}{a value indicating the estimated optimal sparsity parameter, \code{lambda}.}
\item{beta.t.coef}{a length V list of the estimated B-spline coefficient vectors associated with the link functions.}
\item{beta.0.coef}{a length V list of the estimated B-spline coefficient vectors associated with the link functions, used in orthogonalzing the link functions against the main-effect.}
\item{rho.opt}{a length V vector of the estimated optimal penalty parameters used in the P-splines, associated with the V link functions}
\item{link.fn.plot}{a plot object for the (p+q) link functions.}
}
\description{
\code{cmim} is a function for fitting the constrained multiple-index model (CMIM).
Suppose there are V pre-specified groups of scalar-valued covariates for a treatment outcome variable, where each group consists of  multiple (or single) scalar-valued covariates.
\code{cmim}  fits a multiple-index model for estimating treatment-by-covariates interactions, where each index is defined to be a (data-driven) linear combination
of the covariates within group.
The function optimizes the within-group linear combinations, giving V scalar-valued index variables, and simultaneousely
fits a treatment-specific additive model over these scalar-valued index variables (i.e., a multiple-index model).
The differential treatment effect is estimated by these treatment-specific nonparametrically-defined link functions on the indices.
For simultaneous group selection for the treatment effect modifying groups,
 \code{cmim} estimates a sparse combination of the components of the additive model via a \eqn{L1} regularization.
The resulting \code{cmim} object can be used to estimate an optimal treatment selection rule
for a new patient with pre-treatment clinical information.
}
\details{
The sequence of the model coefficients implied by the tuning paramters \code{lambda} is fit by a coordinate descent algorithm.
}
\seealso{
\code{pred.cmim},  \code{cfam},  \code{mim}, \code{fam}
}
\author{
Park, Petkova, Tarpey, Ogden
}
