% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfam-main.R
\name{dataGnFn}
\alias{dataGnFn}
\title{A dataset simulation function}
\usage{
dataGnFn(n = 200, p = 5, q = 5, sigma.eps = 0.4, sigma.u = 0.4,
  correlationZ = 0.1, sigmaZ = pi/2, contrast = "linear", V = 100,
  beta.fn = NULL, sim.set = "A", delta = 1)
}
\arguments{
\item{n}{sample size.}

\item{p}{number of functional-valued pretreatment covariates.}

\item{q}{number of scalar-valued pretreatment covariates.}

\item{sigma.eps}{standard deviation of the random noise term for the outcome model.}

\item{sigma.u}{standard deviation of the measurement error for the functional covariates.}

\item{correlationZ}{correlation among the scalar covariates.}

\item{sigmaZ}{standard deviation of the scalar covariates.}

\item{contrast}{indicates the shape of the treatment-specific link function that defines the interaction effect component.
\describe{
\item{"linear"}{a linear conntrast}
\item{"nonlinear"}{a nonlinear contrast; cosine functions}
}}

\item{V}{the number of support (evaluation) points for functional covariates}

\item{beta.fn}{a list of index coefficient functions associated with the main effect; if \code{NULL}, the coefficients are randomly generated (and scaled to have a unit norm).}

\item{sim.set}{specifies the shape of the index coefficient functions associated with the interaction effect.}

\item{delta}{controls the intensity of the main effect.
\describe{
\item{\code{delta=1}}{moderate main effect}
\item{\code{delta=2}}{big main effect}
}}
}
\value{
\item{y}{a n x 1 vector of treatment outcomes.}
\item{Tr}{a n x 1 vector of treatment indicators.}
\item{X}{a length p list of functional covariates; each element of the p list is a n x V matrix.}
\item{Z}{a n x q matrix of scalar covariates.}
\item{SNR}{the "signal" (interaction effects) to "nuisance" (main effects + noise) variance ratio (SNR) of the dataset.}
\item{optTr}{a n x 1 vector of treatments indicating the optimal treatment selections.}
\item{value.opt}{the "Value" of the optimal treatment selection rule, \code{optTr}.}
}
\description{
\code{dataGenerationFn} generates an example dataset under a model that contains a main effect component, a treatment-by-covariates interaction effect component, and a random noise component.
}
