% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam-main.R
\name{mim}
\alias{mim}
\title{A multiple-index model, for estimating main effects}
\usage{
mim(y, X.list, nbasis = NULL, rho.grid = c(0, 0.25, 0.5),
  lambda.grid = seq(0.1, 1, length.out = 30), type = "AIC",
  coef = NULL, sim.ini = TRUE, cd.eps = 0.001, cd.max.iter = 100,
  linear.link = NULL, eps = 0.01, max.iter = 20, trace = FALSE,
  scale.X = TRUE, sparse.coef = TRUE, plots = TRUE)
}
\arguments{
\item{y}{treatment outcomes, n-by-1 vector}

\item{X.list}{a list of V matrices, in which each matrix is the data matrix corresponding to each of the V groups of the scalar-valued regressors.}

\item{nbasis}{an integer value spefifying the number of B-spline basis funtions for approximating the link function; the default is \code{nbasis=NULL}, and will be determined depending on the sample size.}

\item{rho.grid}{a grid vector of (ridge-type) smoothing parameters the P-splines in the link function estimation}

\item{lambda.grid}{a grid of the sparsity tuning parameters, \code{lambda}, used in the \eqn{L1} regularization.}

\item{type}{can choose bewteen \code{"AIC"}, \code{"BIC"}, and \code{"GCV"}, for the sparsity tuninig parameter selection.}

\item{coef}{a list of length p; can explicitely specify an initial estimate of the index coef vectors; the default is \code{NULL}.}

\item{sim.ini}{if \code{TRUE}, use a sigle-index model estimate as an initial estimate for the index coefficient vectors.}

\item{cd.eps}{a value specifying the convergence criterion of coordinate descent algorithm for estimating the link functions}

\item{cd.max.iter}{an integer value specifying the maximum number of coordinate descent iterations for estimating the link funtions}

\item{linear.link}{a length V vector of 1's or 0's, indicating whether to restrict the jth link functions to be linear or not; if \code{NULL}, will take the vector of 0's.}

\item{eps}{a value specifying the convergence criterion of algorithm for estimating the index coefficient vectors}

\item{max.iter}{an integer value specifying the maximum number of iterations for estimating the index coefficient vectors}

\item{trace}{if \code{TRUE}, show the trace of the fitting procedure; the default is \code{FALSE}.}

\item{scale.X}{if \code{TRUE}, scale the covariates to have a unit variance for the parameter estimation (and will be scaled back for the final estimate)}

\item{sparse.coef}{if \code{TRUE}, estimate spasre index coefficient vectors, using lasso with \code{glmnet}; the default is \code{FALSE}.}

\item{plots}{if \code{TRUE}, produce plots of the estiamated link functions and the estiamated coefficient functions}
}
\value{
a list of information of the fitted constrained multiple-index model including
\item{coef}{a length V list of the estimated index coefficient vectors}
\item{link.fn.fit}{a list of information about the estimated V link functions, including the knot sequences used in the B-spline approximations and the smoothers}
\item{scale.param}{a length V vector representing the scale parameters associated with the V regression component functions; 0 indicates the associated component function is shrinked to 0 (due to the L1 regularization).}
\item{scale.param.path}{a path of \code{scale.param} over a decreasing sequence of the sparsity parameters, \code{lambda}.}
\item{lambda.opt}{a value indicating the estimated optimal sparsity parameter, \code{lambda}.}
\item{beta.t.coef}{a length V list of the estimated B-spline coefficient vectors associated with the link functions.}
\item{beta.0.coef}{a length V list of the estimated B-spline coefficient vectors associated with the link functions, used in orthogonalzing the link functions against the main-effect.}
\item{rho.opt}{a length V vector of the estimated optimal penalty parameters used in the P-splines, associated with the V link functions}
\item{link.fn.plot}{a plot object for the (p+q) link functions.}
}
\description{
\code{mim} is a function for fitting the multiple-index model (MIM).
Suppose there are V pre-specified groups of scalar-valued regressors, in which each group consists of single or multiple scalar-valued regressor(s).
\code{mim} fits a multiple-index model, where each of the indices is defined as a (data-driven) unknown linear combination of the regressors within each group.
\code{mim} optimizes these within-group linear combinations, resulting in V scalar-valued index variables,
and simultaneousely
fits an additive model over these scalar-valued indices (i.e., a multiple-index model).
For simultaneous group selection,
 \code{mim} estimates a sparse combination of the components of this additive model via a \eqn{L1} regularization.
}
\details{
The sequence of the model coefficients implied by the tuning paramters \code{lambda} is fit by a coordinate descent algorithm.
}
\seealso{
\code{pred.cmim},  \code{cfam},  \code{mim}, \code{fam}
}
\author{
Park, Petkova, Tarpey, Ogden
}
