% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfam-main.R
\name{pred.cfam}
\alias{pred.cfam}
\title{\code{cfam} prediction function}
\usage{
pred.cfam(cfam.obj, Xsc = NULL, X1D = NULL)
}
\arguments{
\item{cfam.obj}{a \code{cfam} object}

\item{Xsc}{a matrix of new values for the q pretreatment scalar regressors, at which predictions are to be made.}

\item{X1D}{a list of new values for the p pretreatment functional regressors (each of them evaluated on the same grid as the training data), at which predictions are to be made.}
}
\value{
\item{pred.new}{a matrix of predicted values; column represents each treatment.}
\item{trt.rule}{a vector of suggested treatments}
}
\description{
This function makes predictions from the constrained functional additive model, given a \code{cfam} object and pretreatment covariates, at which predictions are to be made.
The function returns predicted outcomes for each treatment and treatment selection rules.
}
\seealso{
\code{cfam}, \code{cmim}, \code{pred.cmim}
}
\author{
Park, Petkova, Tarpey, Ogden
}
