% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfam-main.R
\name{pred.cmim}
\alias{pred.cmim}
\title{\code{cmim} prediction function}
\usage{
pred.cmim(cmim.obj, X.list = NULL)
}
\arguments{
\item{cmim.obj}{a \code{cmim} object}

\item{X.list}{a list of V matrices, in which each matrix corresponds to new values for each of the V groups of the covariates, at which predictions are to be made.}
}
\value{
\item{pred.new}{a matrix of predicted values; column represents each treatment.}
\item{trt.rule}{a vector of suggested treatments}
}
\description{
This function makes predictions from the constrained multiple-index model, given a \code{cmim} object and pretreatment covariates, at which predictions are to be made.
The function returns predicted outcomes for each treatment and treatment selection rules.
}
\seealso{
\code{cmim}, \code{cfam}, \code{pred.cfam}
}
\author{
Park, Petkova, Tarpey, Ogden
}
