% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam-main.R
\name{pred.fam}
\alias{pred.fam}
\title{\code{fam} prediction function}
\usage{
pred.fam(fam.obj, Xsc = NULL, X1D = NULL)
}
\arguments{
\item{fam.obj}{a \code{fam} object}

\item{Xsc}{a matrix of new values for the q scalar regressors, at which predictions are to be made.}

\item{X1D}{a list of new values for the p functional regressors (each of them evaluated on the same grid as the training data), at which predictions are to be made.}
}
\value{
\item{predicted}{a vector of predicted values.}
}
\description{
This function makes predictions from the functional additive model, given a \code{fam} object and covariates, at which predictions are to be made.
The function returns predicted outcomes.
}
\seealso{
\code{fam}, \code{mim}, \code{pred.mim}
}
\author{
Park, Petkova, Tarpey, Ogden
}
