% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam-main.R
\name{pred.mim}
\alias{pred.mim}
\title{\code{mim} prediction function}
\usage{
pred.mim(mim.obj, X.list = NULL)
}
\arguments{
\item{mim.obj}{a \code{mim} object}

\item{X.list}{a list of V matrices, in which each matrix corresponds to new values for each of the V groups of the covariates, at which predictions are to be made.}
}
\value{
\item{predicted}{a vector of predicted values}
}
\description{
This function makes predictions from the multiple-index model, given a \code{mim} object and pretreatment covariates.
The function returns predicted outcomes.
}
\seealso{
\code{mim}, \code{fam}, \code{pred.fam}
}
\author{
Park, Petkova, Tarpey, Ogden
}
